Refractored the test code and moved to respective folders
