+++ databricks_sql_connector_core/src/databricks_sql_connector_core/sql/thrift_backend.py
@@ -743,7 +743,6 @@ class ThriftBackend:
         else:
             t_result_set_metadata_resp = self._get_metadata_resp(resp.operationHandle)
 
-        # print(f"Line 739 - {t_result_set_metadata_resp.resultFormat}")
         if t_result_set_metadata_resp.resultFormat not in [
             ttypes.TSparkRowSetType.ARROW_BASED_SET,
             ttypes.TSparkRowSetType.COLUMN_BASED_SET,
@@ -880,18 +879,8 @@ class ThriftBackend:
                 # We want to receive proper Timestamp arrow types.
                 "spark.thriftserver.arrowBasedRowSet.timestampAsString": "false"
             },
-            # useArrowNativeTypes=spark_arrow_types,
-            # canReadArrowResult=True,
-            # # canDecompressLZ4Result=lz4_compression,
-            # canDecompressLZ4Result=False,
-            # canDownloadResult=False,
-            # # confOverlay={
-            # #     # We want to receive proper Timestamp arrow types.
-            # #     "spark.thriftserver.arrowBasedRowSet.timestampAsString": "false"
-            # # },
-            # resultDataFormat=TDBSqlResultFormat(None,None,True),
-            # # useArrowNativeTypes=spark_arrow_types,
-             parameters=parameters,
+            useArrowNativeTypes=spark_arrow_types,
+            parameters=parameters,
         )
         resp = self.make_request(self._client.ExecuteStatement, req)
         return self._handle_execute_response(resp, cursor)
