+++ databricks_sql_connector_core/tests/e2e/test_parameterized_queries.py
@@ -8,7 +8,7 @@ from unittest.mock import patch
 import pytest
 import pytz
 
-from databricks.sql import (
+from databricks_sql_connector_core.sql.parameters.native import (
     BigIntegerParameter,
     BooleanParameter,
     DateParameter,
@@ -147,8 +147,8 @@ class TestParameterizedQueries(PySQLPytestTestCase):
         """Applies a patch so we can test the connector's behaviour under different SPARK_CLI_SERVICE_PROTOCOL_VERSION conditions."""
 
         with patch(
-            "databricks.sql.client.Connection.server_parameterized_queries_enabled",
-            return_value=supports_native_params,
+                "databricks_sql_connector_core.sql.client.Connection.server_parameterized_queries_enabled",
+                return_value=supports_native_params,
         ) as mock_parameterized_queries_enabled:
             try:
                 yield mock_parameterized_queries_enabled
@@ -186,10 +186,10 @@ class TestParameterizedQueries(PySQLPytestTestCase):
         return to_return
 
     def _native_roundtrip(
-        self,
-        parameters: Union[Dict, List[Dict]],
-        paramstyle: ParamStyle,
-        parameter_structure: ParameterStructure,
+            self,
+            parameters: Union[Dict, List[Dict]],
+            paramstyle: ParamStyle,
+            parameter_structure: ParameterStructure,
     ):
         if parameter_structure == ParameterStructure.POSITIONAL:
             _query = self.POSITIONAL_PARAMSTYLE_QUERY
@@ -203,11 +203,11 @@ class TestParameterizedQueries(PySQLPytestTestCase):
                 return cursor.fetchone()
 
     def _get_one_result(
-        self,
-        params,
-        approach: ParameterApproach = ParameterApproach.NONE,
-        paramstyle: ParamStyle = ParamStyle.NONE,
-        parameter_structure: ParameterStructure = ParameterStructure.NONE,
+            self,
+            params,
+            approach: ParameterApproach = ParameterApproach.NONE,
+            paramstyle: ParamStyle = ParamStyle.NONE,
+            parameter_structure: ParameterStructure = ParameterStructure.NONE,
     ):
         """When approach is INLINE then we use %(param)s paramstyle and a connection with use_inline_params=True
         When approach is NATIVE then we use :param paramstyle and a connection with use_inline_params=False
@@ -243,12 +243,12 @@ class TestParameterizedQueries(PySQLPytestTestCase):
         "approach,paramstyle,parameter_structure", approach_paramstyle_combinations
     )
     def test_primitive_single(
-        self,
-        approach,
-        paramstyle,
-        parameter_structure,
-        primitive: Primitive,
-        inline_table,
+            self,
+            approach,
+            paramstyle,
+            parameter_structure,
+            primitive: Primitive,
+            inline_table,
     ):
         """When ParameterApproach.INLINE is passed, inferrence will not be used.
         When ParameterApproach.NATIVE is passed, primitive inputs will be inferred.
@@ -285,10 +285,10 @@ class TestParameterizedQueries(PySQLPytestTestCase):
         ],
     )
     def test_dbsqlparameter_single(
-        self,
-        primitive: Primitive,
-        dbsql_parameter_cls: Type[TDbsqlParameter],
-        parameter_structure: ParameterStructure,
+            self,
+            primitive: Primitive,
+            dbsql_parameter_cls: Type[TDbsqlParameter],
+            parameter_structure: ParameterStructure,
     ):
         dbsql_param = dbsql_parameter_cls(
             value=primitive.value,  # type: ignore
@@ -316,11 +316,11 @@ class TestParameterizedQueries(PySQLPytestTestCase):
                     cursor.execute("SELECT %(p)s", parameters={"p": 1})
                     if use_inline_params is True:
                         assert (
-                            "Consider using native parameters." in caplog.text
+                                "Consider using native parameters." in caplog.text
                         ), "Log message should be suppressed"
                     elif use_inline_params == "silent":
                         assert (
-                            "Consider using native parameters." not in caplog.text
+                                "Consider using native parameters." not in caplog.text
                         ), "Log message should not be supressed"
 
     def test_positional_native_params_with_defaults(self):
@@ -333,12 +333,12 @@ class TestParameterizedQueries(PySQLPytestTestCase):
     @pytest.mark.parametrize(
         "params",
         (
-            [
-                StringParameter(value="foo"),
-                StringParameter(value="bar"),
-                StringParameter(value="baz"),
-            ],
-            ["foo", "bar", "baz"],
+                [
+                    StringParameter(value="foo"),
+                    StringParameter(value="bar"),
+                    StringParameter(value="baz"),
+                ],
+                ["foo", "bar", "baz"],
         ),
     )
     def test_positional_native_multiple(self, params):
