+++ databricks_sql_connector_core/tests/unit/test_cloud_fetch_queue.py
@@ -3,8 +3,8 @@ import unittest
 from unittest.mock import MagicMock, patch
 from ssl import create_default_context
 
-from databricks.sql import TSparkArrowResultLink
-import databricks.sql.utils as utils
+from databricks_sql_connector_core.sql.thrift_api.TCLIService.ttypes import TSparkArrowResultLink
+import databricks_sql_connector_core.sql.utils as utils
 
 class CloudFetchQueueSuite(unittest.TestCase):
 
@@ -43,7 +43,7 @@ class CloudFetchQueueSuite(unittest.TestCase):
         return sink.getvalue().to_pybytes()
 
 
-    @patch("databricks.sql.utils.CloudFetchQueue._create_next_table", return_value=[None, None])
+    @patch("databricks_sql_connector_core.sql.utils.CloudFetchQueue._create_next_table", return_value=[None, None])
     def test_initializer_adds_links(self, mock_create_next_table):
         schema_bytes = MagicMock()
         result_links = self.create_result_links(10)
@@ -72,7 +72,7 @@ class CloudFetchQueueSuite(unittest.TestCase):
         assert len(queue.download_manager._download_tasks) == 0
         assert queue.table is None
 
-    @patch("databricks.sql.cloudfetch.download_manager.ResultFileDownloadManager.get_next_downloaded_file", return_value=None)
+    @patch("databricks_sql_connector_core.sql.cloudfetch.download_manager.ResultFileDownloadManager.get_next_downloaded_file", return_value=None)
     def test_create_next_table_no_download(self, mock_get_next_downloaded_file):
         queue = utils.CloudFetchQueue(
             MagicMock(),
@@ -84,8 +84,8 @@ class CloudFetchQueueSuite(unittest.TestCase):
         assert queue._create_next_table() is None
         mock_get_next_downloaded_file.assert_called_with(0)
 
-    @patch("databricks.sql.utils.create_arrow_table_from_arrow_file")
-    @patch("databricks.sql.cloudfetch.download_manager.ResultFileDownloadManager.get_next_downloaded_file",
+    @patch("databricks_sql_connector_core.sql.utils.create_arrow_table_from_arrow_file")
+    @patch("databricks_sql_connector_core.sql.cloudfetch.download_manager.ResultFileDownloadManager.get_next_downloaded_file",
            return_value=MagicMock(file_bytes=b"1234567890", row_count=4))
     def test_initializer_create_next_table_success(self, mock_get_next_downloaded_file, mock_create_arrow_table):
         mock_create_arrow_table.return_value = self.make_arrow_table()
@@ -111,7 +111,7 @@ class CloudFetchQueueSuite(unittest.TestCase):
         assert table.num_rows == 4
         assert queue.start_row_index == 8
 
-    @patch("databricks.sql.utils.CloudFetchQueue._create_next_table")
+    @patch("databricks_sql_connector_core.sql.utils.CloudFetchQueue._create_next_table")
     def test_next_n_rows_0_rows(self, mock_create_next_table):
         mock_create_next_table.return_value = self.make_arrow_table()
         schema_bytes, description = MagicMock(), MagicMock()
@@ -131,7 +131,7 @@ class CloudFetchQueueSuite(unittest.TestCase):
         assert queue.table_row_index == 0
         assert result == self.make_arrow_table()[0:0]
 
-    @patch("databricks.sql.utils.CloudFetchQueue._create_next_table")
+    @patch("databricks_sql_connector_core.sql.utils.CloudFetchQueue._create_next_table")
     def test_next_n_rows_partial_table(self, mock_create_next_table):
         mock_create_next_table.return_value = self.make_arrow_table()
         schema_bytes, description = MagicMock(), MagicMock()
@@ -151,7 +151,7 @@ class CloudFetchQueueSuite(unittest.TestCase):
         assert queue.table_row_index == 3
         assert result == self.make_arrow_table()[:3]
 
-    @patch("databricks.sql.utils.CloudFetchQueue._create_next_table")
+    @patch("databricks_sql_connector_core.sql.utils.CloudFetchQueue._create_next_table")
     def test_next_n_rows_more_than_one_table(self, mock_create_next_table):
         mock_create_next_table.return_value = self.make_arrow_table()
         schema_bytes, description = MagicMock(), MagicMock()
@@ -171,7 +171,7 @@ class CloudFetchQueueSuite(unittest.TestCase):
         assert queue.table_row_index == 3
         assert result == pyarrow.concat_tables([self.make_arrow_table(), self.make_arrow_table()])[:7]
 
-    @patch("databricks.sql.utils.CloudFetchQueue._create_next_table")
+    @patch("databricks_sql_connector_core.sql.utils.CloudFetchQueue._create_next_table")
     def test_next_n_rows_only_one_table_returned(self, mock_create_next_table):
         mock_create_next_table.side_effect = [self.make_arrow_table(), None]
         schema_bytes, description = MagicMock(), MagicMock()
@@ -190,7 +190,7 @@ class CloudFetchQueueSuite(unittest.TestCase):
         assert result.num_rows == 4
         assert result == self.make_arrow_table()
 
-    @patch("databricks.sql.utils.CloudFetchQueue._create_next_table", return_value=None)
+    @patch("databricks_sql_connector_core.sql.utils.CloudFetchQueue._create_next_table", return_value=None)
     def test_next_n_rows_empty_table(self, mock_create_next_table):
         schema_bytes = self.get_schema_bytes()
         description = MagicMock()
@@ -207,7 +207,7 @@ class CloudFetchQueueSuite(unittest.TestCase):
         mock_create_next_table.assert_called()
         assert result == pyarrow.ipc.open_stream(bytearray(schema_bytes)).read_all()
 
-    @patch("databricks.sql.utils.CloudFetchQueue._create_next_table")
+    @patch("databricks_sql_connector_core.sql.utils.CloudFetchQueue._create_next_table")
     def test_remaining_rows_empty_table_fully_returned(self, mock_create_next_table):
         mock_create_next_table.side_effect = [self.make_arrow_table(), None, 0]
         schema_bytes, description = MagicMock(), MagicMock()
@@ -226,7 +226,7 @@ class CloudFetchQueueSuite(unittest.TestCase):
         assert result.num_rows == 0
         assert result == self.make_arrow_table()[0:0]
 
-    @patch("databricks.sql.utils.CloudFetchQueue._create_next_table")
+    @patch("databricks_sql_connector_core.sql.utils.CloudFetchQueue._create_next_table")
     def test_remaining_rows_partial_table_fully_returned(self, mock_create_next_table):
         mock_create_next_table.side_effect = [self.make_arrow_table(), None]
         schema_bytes, description = MagicMock(), MagicMock()
@@ -245,7 +245,7 @@ class CloudFetchQueueSuite(unittest.TestCase):
         assert result.num_rows == 2
         assert result == self.make_arrow_table()[2:]
 
-    @patch("databricks.sql.utils.CloudFetchQueue._create_next_table")
+    @patch("databricks_sql_connector_core.sql.utils.CloudFetchQueue._create_next_table")
     def test_remaining_rows_one_table_fully_returned(self, mock_create_next_table):
         mock_create_next_table.side_effect = [self.make_arrow_table(), None]
         schema_bytes, description = MagicMock(), MagicMock()
@@ -264,7 +264,7 @@ class CloudFetchQueueSuite(unittest.TestCase):
         assert result.num_rows == 4
         assert result == self.make_arrow_table()
 
-    @patch("databricks.sql.utils.CloudFetchQueue._create_next_table")
+    @patch("databricks_sql_connector_core.sql.utils.CloudFetchQueue._create_next_table")
     def test_remaining_rows_multiple_tables_fully_returned(self, mock_create_next_table):
         mock_create_next_table.side_effect = [self.make_arrow_table(), self.make_arrow_table(), None]
         schema_bytes, description = MagicMock(), MagicMock()
@@ -284,7 +284,7 @@ class CloudFetchQueueSuite(unittest.TestCase):
         assert result.num_rows == 5
         assert result == pyarrow.concat_tables([self.make_arrow_table(), self.make_arrow_table()])[3:]
 
-    @patch("databricks.sql.utils.CloudFetchQueue._create_next_table", return_value=None)
+    @patch("databricks_sql_connector_core.sql.utils.CloudFetchQueue._create_next_table", return_value=None)
     def test_remaining_rows_empty_table(self, mock_create_next_table):
         schema_bytes = self.get_schema_bytes()
         description = MagicMock()
