+++ databricks_sql_connector_core/tests/unit/test_fetches.py
@@ -3,8 +3,8 @@ from unittest.mock import Mock
 
 import pyarrow as pa
 
-import databricks.sql.client as client
-from databricks.sql import ExecuteResponse, ArrowQueue
+import databricks_sql_connector_core.sql.client as client
+from databricks_sql_connector_core.sql.utils import ExecuteResponse, ArrowQueue
 
 
 class FetchTests(unittest.TestCase):
