+++ databricks_sql_connector_core/tests/unit/test_retry.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock, patch
 import pytest
 from requests import Request
 from urllib3 import HTTPResponse
-from databricks.sql import DatabricksRetryPolicy, RequestHistory
+from databricks_sql_connector_core.sql.auth.retry import DatabricksRetryPolicy, RequestHistory
 
 
 class TestRetry:
