+++ examples/interactive_oauth.py
@@ -1,4 +1,4 @@
-from databricks import sql
+from databricks_sql_connector_core import sql
 import os
 
 """databricks-sql-connector supports user to machine OAuth login which means the
@@ -14,8 +14,7 @@ token across script executions.
 """
 
 with sql.connect(server_hostname = os.getenv("DATABRICKS_SERVER_HOSTNAME"),
-                 http_path       = os.getenv("DATABRICKS_HTTP_PATH"),
-                 auth_type="databricks-oauth") as connection:
+                 http_path       = os.getenv("DATABRICKS_HTTP_PATH")) as connection:
 
     for x in range(1, 100):
         cursor = connection.cursor()
