+++ examples/persistent_oauth.py
@@ -16,29 +16,29 @@ For this example, the DevOnlyFilePersistence class is provided. Do not use this 
 import os
 from typing import Optional
 
-from databricks import sql
-from databricks.sql import OAuthPersistence, OAuthToken, DevOnlyFilePersistence
+from databricks_sql_connector_core import sql
+from databricks_sql_connector_core.sql.experimental.oauth_persistence import OAuthPersistence, OAuthToken, DevOnlyFilePersistence
 
 
 class SampleOAuthPersistence(OAuthPersistence):
-  def persist(self, hostname: str, oauth_token: OAuthToken):
-    """To be implemented by the end user to persist in the preferred storage medium.
+    def persist(self, hostname: str, oauth_token: OAuthToken):
+        """To be implemented by the end user to persist in the preferred storage medium.
 
-    OAuthToken has two properties:
-        1. OAuthToken.access_token
-        2. OAuthToken.refresh_token
+        OAuthToken has two properties:
+            1. OAuthToken.access_token
+            2. OAuthToken.refresh_token
 
-    Both should be persisted.
-    """
-    pass
+        Both should be persisted.
+        """
+        pass
 
-  def read(self, hostname: str) -> Optional[OAuthToken]:
-    """To be implemented by the end user to fetch token from the preferred storage
+    def read(self, hostname: str) -> Optional[OAuthToken]:
+        """To be implemented by the end user to fetch token from the preferred storage
 
-    Fetch the access_token and refresh_token for the given hostname.
-    Return OAuthToken(access_token, refresh_token)
-    """
-    pass
+        Fetch the access_token and refresh_token for the given hostname.
+        Return OAuthToken(access_token, refresh_token)
+        """
+        pass
 
 with sql.connect(server_hostname = os.getenv("DATABRICKS_SERVER_HOSTNAME"),
                  http_path       = os.getenv("DATABRICKS_HTTP_PATH"),
